pyproject.toml
[project]
name = "doc-arag"
version = "0.1.0"
description = "Agentic RAG system with document parsing and vector search"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.6.10",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "weaviate-client>=4.9.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "minio>=7.2.0",
    "python-multipart>=0.0.20",
    "grpcio>=1.60.0",
    "langchain-text-splitters>=0.3.11",
    "tenacity>=9.1.2",
    "pypdf==6.1.3",
    "python-magic>=0.4.27",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "pysentry-rs>=0.3.12",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.25.2",
    "ruff>=0.14.0",
]

[tool.ruff]
exclude = ["*_pb2.py", "*_pb2_grpc.py"]

[tool.mypy]
mypy_path = "."
explicit_package_bases = true
namespace_packages = true
exclude = [".*_pb2.py$", ".*_pb2_grpc.py$"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
